+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.12.1 - Unreleased
 
+* Fix for pipeline export with multiple materials without name, #144
+
 ### 0.12.0 (2019-Oct-02)
 
 * Added support for `ignore_for_scheduling` to dependency materials. Updated README with changes in new Format Versions.